feat: Use Double for event ratings and adjust related DTOs
This commit refactors the event rating representation to use `Double`
instead of `BigDecimal` for improved consistency and simpler calculations.
It also adjusts related DTOs and repositories to reflect this change.

Key changes:
- Changed `EventRatingCount` to use `Double` for the rating.
- Modified `EventFullDto` and `EventShortDto` to use `Double` for the
  rating field.
- Updated `EventReviewRepository` to use `AVG(er.score)` which returns
  a `Double` directly, removing the need for casting.
- Adjusted `EventStatsCollectorImpl` to use `Double` for rating maps
  and default values.
- Updated `UserReviewRepository` to ensure `existsByParams` method
  signature is correct.

+++ main-service/src/main/java/ru/practicum/main/event/dto/EventRatingCount.java
@@ -1,6 +1,4 @@
 package ru.practicum.main.event.dto;
 
-import java.math.BigDecimal;
-
-public record EventRatingCount(Long eventId, BigDecimal rating) {
+public record EventRatingCount(Long eventId, Double rating) {
 }
