+++ main-service/src/main/java/ru/practicum/main/event/dto/mapper/EventMapper.java
@@ -16,7 +16,6 @@ import ru.practicum.main.event.model.Event;
 import ru.practicum.main.event.model.Location;
 import ru.practicum.main.user.model.User;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Mapper(componentModel = "spring", imports = LocalDateTime.class)
@@ -30,9 +29,9 @@ public interface EventMapper {
 
     Location toLocation(LocationDto locationDto);
 
-    EventFullDto toFullDto(Event event, Long views, Long confirmedRequests, BigDecimal rating);
+    EventFullDto toFullDto(Event event, Long views, Long confirmedRequests, Double rating);
 
-    EventShortDto toShortDto(Event event, Long views, Long confirmedRequests, BigDecimal rating);
+    EventShortDto toShortDto(Event event, Long views, Long confirmedRequests, Double rating);
 
     @BeanMapping(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE)
     @Mapping(target = "id", ignore = true)
